Flatten out source and sink
